ApollosApplication-master/db/requirements.txt
boto3
flask
awscli
flask-bcrypt
pyOpenSSL
PyJWT
datetime
os
shutil
sys
dynamodb_encryption_sdk
